pyproject.toml
[project]
name = "merit-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "py-cord>=2.6.1",
    "python-dotenv>=1.0.1",
    "tinydb>=4.8.2",
]
